requirements.txt
pip~=25.1.1
attrs~=25.3.0
filelock~=3.18.0
redis~=6.4.0
setuptools~=65.5.0
packaging~=25.0
typing_extensions~=4.14.0
requests~=2.32.4
pandas~=2.3.2
tqdm~=4.67.1
python-dotenv~=1.1.1